feat: Enhance printer discovery and management features
- Updated EnhancedPrinterCard component to dynamically construct server URLs based on frontend ports.
- Modified PrinterDiscoveryModal to include backend and frontend ports, as well as camera stream and snapshot URLs in the discovered printer DTO.
- Extended DiscoveredPrinterDto to include optional backend and frontend ports, along with camera URLs.
- Improved PrintersController to handle discovery session cancellation without async token.
- Fixed EfSliceJobRepository to correctly sum artifact sizes.
- Refactored TokenRevocationCleanupService and WorkerHealthMonitorService to remove unnecessary cancellation tokens.
- Enhanced MoonrakerDiscoveryProbe to discover frontend ports and camera URLs.
- Updated IMoonrakerClient interface and its implementation to support querying for camera URLs.
- Adjusted PrintersService to utilize discovered backend and frontend ports, and camera URLs.
- Added utility methods for stripping ports from server URLs in various services.
- Improved test factory to ensure test artifacts directory management is robust.
- Refined HttpJobPollerService to use async cancellation for better resource management.

+++ src/api/Repositories/Slicing/EfSliceJobRepository.cs
@@ -161,7 +161,7 @@ public class EfSliceJobRepository : ISliceJobRepository
         // Aggregate bytes from artifacts table
         if (ids.Length > 0)
         {
-            long totalBytes = await _db.Artifacts.Where(a => ids.Contains(a.Id)).SumAsync(a => (long)a.SizeBytes, ct);
+            long totalBytes = await _db.Artifacts.Where(a => ids.Contains(a.Id)).SumAsync(a => a.SizeBytes, ct);
             job.ArtifactsTotalBytes = totalBytes;
             job.ArtifactsCount = ids.Length;
         }
